pyproject.toml
[project]
name="reddit-meter-backend"
version="0.1.0"
description="backend for my Reddit Meter backend project"
requires-python=">=3.10"

dependencies = [
    "black>=25.1.0",
    "datasets>=4.0.0",
    "fastapi[all]>=0.115.13",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-storage>=3.2.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.3",
    "memory-profiler>=0.61.0",
    "pandas>=2.3.0",
    "praw>=7.8.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "ruff>=0.12.0",
    "secure>=1.0.1",
    "slowapi>=0.1.9",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "huggingface-hub>=0.33.1",
    "matplotlib>=3.10.3",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
]
